+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.0 – 2026-03-21
+
+- Added current temperature.
+- Changed layout to table format.
+- Added support for Pebble Round 2.
+- Changed font to Inter.
+
 ## 0.1.0 – 2026-01-17
 
 - First release.